Transaction as interface #1053
